Use new custom dropdown widget in mode picker
